Final Project/dashboard/requirements.txt
flask==3.0.3
paho-mqtt==2.1.0
matplotlib==3.9.2